rename commands file removing some old files
