+++ src/pieces/command_interface/open_command.py
@@ -59,4 +59,3 @@ class OpenCommand(BaseCommand):
         """Execute the open command."""
         open_command(**kwargs)
         return 0
-
