+++ src/pieces/command_interface/simple_commands.py
@@ -8,7 +8,10 @@ from pieces.core import (
     onboarding_command,
     PiecesInstaller,
 )
-from pieces.core.execute_command import ExecuteCommand as OldExecuteCommand
+from pieces.core.execute_command import ExecuteCommand as CoreExecute
+from pieces.gui import print_version_details
+from pieces import __version__
+from pieces.settings import Settings
 
 
 class RunCommand(BaseCommand):
@@ -59,7 +62,7 @@ class ExecuteCommand(BaseCommand):
         pass
 
     def execute(self, **kwargs) -> int:
-        OldExecuteCommand.execute_command(**kwargs)
+        CoreExecute.execute_command(**kwargs)
         return 0
 
 
@@ -166,3 +169,33 @@ class OnboardingCommand(BaseCommand):
         onboarding_command(**kwargs)
         return 0
 
+
+class VersionCommand(BaseCommand):
+    """Command to display version information for Pieces CLI and PiecesOS."""
+
+    def get_name(self) -> str:
+        return "version"
+
+    def get_help(self) -> str:
+        return "Gets version of PiecesOS"
+
+    def get_description(self) -> str:
+        return "Display version information for both Pieces CLI and PiecesOS, including build numbers and compatibility details"
+
+    def get_examples(self) -> list[str]:
+        return ["pieces version"]
+
+    def get_docs(self) -> str:
+        return URLs.CLI_VERSION_DOCS.value
+
+    def add_arguments(self, parser: argparse.ArgumentParser):
+        """Version command has no additional arguments."""
+        pass
+
+    def execute(self, **kwargs) -> int:
+        """Execute the version command."""
+        if Settings.pieces_os_version:
+            print_version_details(Settings.pieces_os_version, __version__)
+        else:
+            pass
+        return 0
