+++ src/pieces/core/__init__.py
@@ -2,7 +2,6 @@ from .cli_loop import loop
 from .change_model import change_model
 from .search_command import search
 from .list_command import ListCommand
-from .version_command import version
 from .auth_commands import sign_in, sign_out
 from .config_command import ConfigCommands
 from .execute_command import ExecuteCommand
@@ -14,7 +13,6 @@ from .open_command import open_command
 
 __all__ = [
     "loop",
-    "version",
     "search",
     "change_model",
     "sign_out",
